move migrattion scripts
