+++ src/migration/PoC_dfusion.js
@@ -1,4 +1,4 @@
-const { isDevelopmentNetwork, getDependency } = require("./migration_utilities.js")
+const { isDevelopmentNetwork, getDependency } = require("./utilities.js")
 const deployOwl = require("@gnosis.pm/owl-token/src/migrations-truffle-5/3_deploy_OWL")
 
 async function migrate({ artifacts, deployer, network, accounts, web3, feeDenominator = 1000, maxTokens = 2 ** 16 - 1 }) {
